refactor: add dedicated tuple namespace
