+++ include/Simple-Utility/functional/operators/bind.hpp
@@ -8,9 +8,8 @@
 
 #pragma once
 
-#include "Simple-Utility/tuple_utility.hpp"
-#include "Simple-Utility/concepts/utility.hpp"
 #include "Simple-Utility/functional/base.hpp"
+#include "Simple-Utility/tuple/general.hpp"
 
 namespace sl::functional::operators::detail
 {
@@ -18,17 +17,15 @@ namespace sl::functional::operators::detail
 	{
 		template <class TCallArgsTuple, class TFunction, class... TBoundFunctions>
 			requires concepts::apply_invocable<
-				TFunction, tuple_cat_result_t<std::tuple<std::invoke_result_t<TBoundFunctions>...>, TCallArgsTuple>
+				TFunction, tuple::tuple_cat_result_t<std::tuple<std::invoke_result_t<TBoundFunctions>...>, TCallArgsTuple>
 			>
 		[[nodiscard]]
-		constexpr auto operator ()
-		(
+		constexpr auto operator ()(
 			TCallArgsTuple&& callArgsTuple,
 			TFunction&& func,
 			TBoundFunctions&&... boundFunctions
-		) const
-		noexcept(concepts::nothrow_apply_invocable<
-			TFunction, tuple_cat_result_t<std::tuple<std::invoke_result_t<TBoundFunctions>...>, TCallArgsTuple>
+		) const noexcept(concepts::nothrow_apply_invocable<
+			TFunction, tuple::tuple_cat_result_t<std::tuple<std::invoke_result_t<TBoundFunctions>...>, TCallArgsTuple>
 		>)
 		{
 			return std::apply(
@@ -52,17 +49,15 @@ namespace sl::functional::operators::detail
 	{
 		template <class TCallArgsTuple, class TFunction, class... TBoundFunctions>
 			requires concepts::apply_invocable<
-				TFunction, tuple_cat_result_t<TCallArgsTuple, std::tuple<std::invoke_result_t<TBoundFunctions>...>>
+				TFunction, tuple::tuple_cat_result_t<TCallArgsTuple, std::tuple<std::invoke_result_t<TBoundFunctions>...>>
 			>
 		[[nodiscard]]
-		constexpr auto operator ()
-		(
+		constexpr auto operator ()(
 			TCallArgsTuple&& callArgsTuple,
 			TFunction&& func,
 			TBoundFunctions&&... boundFunctions
-		) const
-		noexcept(concepts::nothrow_apply_invocable<
-			TFunction, tuple_cat_result_t<TCallArgsTuple, std::tuple<std::invoke_result_t<TBoundFunctions>...>>
+		) const noexcept(concepts::nothrow_apply_invocable<
+			TFunction, tuple::tuple_cat_result_t<TCallArgsTuple, std::tuple<std::invoke_result_t<TBoundFunctions>...>>
 		>)
 		{
 			return std::apply(
@@ -96,7 +91,8 @@ namespace sl::functional::operators
 	 * \relatesalso sl::functional::enable_operation
 	 */
 	struct bind_front
-	{};
+	{
+	};
 
 	/**
 	 * \brief Specialized traits for \ref bind_front.
@@ -106,7 +102,7 @@ namespace sl::functional::operators
 	struct tag_traits<bind_front>
 	{
 		using operation_t = detail::bind_front_caller_fn;
-		static constexpr composition_strategy_t composition_strategy{ composition_strategy_t::join };
+		static constexpr composition_strategy_t composition_strategy{composition_strategy_t::join};
 	};
 
 	/**
@@ -121,16 +117,14 @@ namespace sl::functional::operators
 	template <derived_from_unified_base<functional::detail::enable_operators_base_tag> TFunc, class TValue>
 		requires std::derived_from<std::remove_cvref_t<TFunc>, bind_front>
 	[[nodiscard]]
-	constexpr auto operator <<
-	(
+	constexpr auto operator <<(
 		TFunc&& func,
 		TValue&& value
-	)
-	noexcept(is_nothrow_composable_v<bind_front, TFunc, value_fn<std::decay_t<TValue>>&&>)
+	) noexcept(is_nothrow_composable_v<bind_front, TFunc, value_fn<std::decay_t<TValue>>&&>)
 	{
 		return functional::detail::make_composition_from_tag<bind_front>(
 			std::forward<TFunc>(func),
-			value_fn{ std::forward<TValue>(value) }
+			value_fn{std::forward<TValue>(value)}
 		);
 	}
 
@@ -139,7 +133,8 @@ namespace sl::functional::operators
 	 * \relatesalso sl::functional::enable_operation
 	 */
 	struct bind_back
-	{};
+	{
+	};
 
 	/**
 	 * \brief Specialized traits for \ref bind_back.
@@ -149,7 +144,7 @@ namespace sl::functional::operators
 	struct tag_traits<bind_back>
 	{
 		using operation_t = detail::bind_back_caller_fn;
-		static constexpr composition_strategy_t composition_strategy{ composition_strategy_t::join };
+		static constexpr composition_strategy_t composition_strategy{composition_strategy_t::join};
 	};
 
 	/**
@@ -164,16 +159,14 @@ namespace sl::functional::operators
 	template <derived_from_unified_base<functional::detail::enable_operators_base_tag> TFunc, class TValue>
 		requires std::derived_from<std::remove_cvref_t<TFunc>, bind_back>
 	[[nodiscard]]
-	constexpr auto operator >>
-	(
+	constexpr auto operator >>(
 		TFunc&& func,
 		TValue&& value
-	)
-	noexcept(is_nothrow_composable_v<bind_back, TFunc, value_fn<std::decay_t<TValue>>&&>)
+	) noexcept(is_nothrow_composable_v<bind_back, TFunc, value_fn<std::decay_t<TValue>>&&>)
 	{
 		return functional::detail::make_composition_from_tag<bind_back>(
 			std::forward<TFunc>(func),
-			value_fn{ std::forward<TValue>(value) }
+			value_fn{std::forward<TValue>(value)}
 		);
 	}
