+++ include/Simple-Utility/functional/operators/pipe.hpp
@@ -8,9 +8,8 @@
 
 #pragma once
 
-#include "Simple-Utility/tuple_utility.hpp"
-#include "Simple-Utility/concepts/utility.hpp"
 #include "Simple-Utility/functional/base.hpp"
+#include "Simple-Utility/tuple/general.hpp"
 
 namespace sl::functional::operators::detail
 {
@@ -20,12 +19,14 @@ namespace sl::functional::operators::detail
 			std::is_invocable_v<TFunc, TInput>
 			&& is_recursive_supply_invocable<std::invoke_result_t<TFunc, TInput>, TOthers...>::value
 		>
-	{ };
+	{
+	};
 
 	template <class TInput, std::invocable<TInput> TFunc>
 	struct is_recursive_supply_invocable<TInput, TFunc>
 		: std::bool_constant<std::is_invocable_v<TFunc, TInput>>
-	{ };
+	{
+	};
 
 	template <class TInput, class... TFunctions>
 	concept recursive_supply_invocable = is_recursive_supply_invocable<TInput, TFunctions...>::value;
@@ -36,12 +37,14 @@ namespace sl::functional::operators::detail
 			std::is_nothrow_invocable_v<TFunc, TInput>
 			&& is_nothrow_recursive_supply_invocable<std::invoke_result_t<TFunc, TInput>, TOthers...>::value
 		>
-	{ };
+	{
+	};
 
 	template <class TInput, std::invocable<TInput> TFunc>
 	struct is_nothrow_recursive_supply_invocable<TInput, TFunc>
 		: std::bool_constant<std::is_nothrow_invocable_v<TFunc, TInput>>
-	{ };
+	{
+	};
 
 	template <class TInput, class... TFunctions>
 	concept nothrow_recursive_supply_invocable = is_nothrow_recursive_supply_invocable<TInput, TFunctions...>::value;
@@ -49,13 +52,11 @@ namespace sl::functional::operators::detail
 	template <class TInput, std::invocable<TInput> TFunc, class... TOthers>
 		requires recursive_supply_invocable<TInput, TFunc, TOthers...>
 	[[nodiscard]]
-	constexpr decltype(auto) recursive_supply_invoke
-	(
+	constexpr decltype(auto) recursive_supply_invoke(
 		TInput&& input,
 		TFunc&& func,
 		TOthers&&... otherFunctions
-	)
-	noexcept(nothrow_recursive_supply_invocable<TInput, TFunc, TOthers...>)
+	) noexcept(nothrow_recursive_supply_invocable<TInput, TFunc, TOthers...>)
 	{
 		if constexpr (0 < sizeof...(TOthers))
 		{
@@ -73,16 +74,15 @@ namespace sl::functional::operators::detail
 	struct nested_invoke_caller_fn
 	{
 		template <class TCallArgsTuple, concepts::apply_invocable<TCallArgsTuple> TInitFunction, class... TFunctions>
-			requires recursive_supply_invocable<apply_invoke_result_t<TInitFunction, TCallArgsTuple>, TFunctions...>
+			requires recursive_supply_invocable<tuple::apply_invoke_result_t<TInitFunction, TCallArgsTuple>, TFunctions...>
 		[[nodiscard]]
-		constexpr auto operator ()
-		(
+		constexpr auto operator ()(
 			TCallArgsTuple&& callArgsTuple,
 			TInitFunction&& initFunction,
 			TFunctions&&... functions
-		) const
-		noexcept(concepts::nothrow_apply_invocable<TInitFunction, TCallArgsTuple>
-				&& nothrow_recursive_supply_invocable<apply_invoke_result_t<TInitFunction, TCallArgsTuple>, TFunctions...>)
+		) const noexcept(concepts::nothrow_apply_invocable<TInitFunction, TCallArgsTuple>
+						&& nothrow_recursive_supply_invocable<
+							tuple::apply_invoke_result_t<TInitFunction, TCallArgsTuple>, TFunctions...>)
 		{
 			return recursive_supply_invoke(
 				std::apply(std::forward<TInitFunction>(initFunction), std::forward<TCallArgsTuple>(callArgsTuple)),
@@ -105,7 +105,8 @@ namespace sl::functional::operators
 	 * \relatesalso sl::functional::enable_operation
 	 */
 	struct pipe
-	{};
+	{
+	};
 
 	/**
 	 * \brief Specialized traits for \ref pipe.
@@ -115,7 +116,7 @@ namespace sl::functional::operators
 	struct tag_traits<pipe>
 	{
 		using operation_t = detail::nested_invoke_caller_fn;
-		static constexpr composition_strategy_t composition_strategy{ composition_strategy_t::join };
+		static constexpr composition_strategy_t composition_strategy{composition_strategy_t::join};
 	};
 
 	/**
@@ -133,12 +134,10 @@ namespace sl::functional::operators
 				|| (std::derived_from<std::remove_cvref_t<TRhs>, pipe>
 					&& derived_from_unified_base<TRhs, functional::detail::enable_operators_base_tag>)
 	[[nodiscard]]
-	constexpr auto operator |
-	(
+	constexpr auto operator |(
 		TLhs&& lhs,
 		TRhs&& rhs
-	)
-	noexcept(is_nothrow_composable_v<pipe, TLhs, TRhs>)
+	) noexcept(is_nothrow_composable_v<pipe, TLhs, TRhs>)
 	{
 		return functional::detail::make_composition_from_tag<pipe>(std::forward<TLhs>(lhs), std::forward<TRhs>(rhs));
 	}
