+++ include/Simple-Utility/tuple.hpp
@@ -1,0 +1,14 @@
+//          Copyright Dominic Koepke 2019 - 2023.
+// Distributed under the Boost Software License, Version 1.0.
+//    (See accompanying file LICENSE_1_0.txt or copy at
+//          https://www.boost.org/LICENSE_1_0.txt)
+
+#ifndef SL_UTILITY_TUPLE_HPP
+#define SL_UTILITY_TUPLE_HPP
+
+#pragma once
+
+#include "Simple-Utility/tuple/algorithm.hpp"
+#include "Simple-Utility/tuple/general.hpp"
+
+#endif
