+++ tests/CMakeLists.txt
@@ -7,7 +7,7 @@ add_executable(
 	"helper.cpp"
 	"math.cpp"
 	"unique_handle.cpp"
-	"tuple_utility.cpp"
+	"tuple.cpp"
 )
 
 add_subdirectory("nullables")
